problem with iterating over array_of_jobs
