+++ lib/remote_jobs.rb
@@ -6,7 +6,7 @@ require_relative './remote_jobs/version.rb'
 require_relative './remote_jobs/cli'
 require_relative './remote_jobs/scraper'
 require_relative './remote_jobs/details'
-require_relative './remote_jobs/jobs'
+#require_relative './remote_jobs/jobs'
 
 #module RemoteJobs
   # Your code goes here...
